isolate tool definitions & registration
